% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_createGene.R
\name{createGene}
\alias{createGene}
\title{Create a Gene object}
\usage{
createGene(ID, symbol, name, description, chr, start, end, strand, product)
}
\arguments{
\item{ID}{integer The gene ID (e.g., Ensembl ID or NCBI gene ID).}

\item{symbol}{character. The gene symbol.}

\item{name}{character. The gene name.}

\item{description}{character. The gene description.}

\item{chr}{character. The chromosome location.}

\item{start}{numeric. The start position of the gene.}

\item{end}{numeric. The end position of the gene.}

\item{strand}{character. The strand of the gene.}

\item{product}{list. The gene product.}
}
\value{
A Gene object.
}
\description{
This function creates a new Gene object.
}
\examples{
gene <- createGene(1L, "SYMBOL", "Gene Name", "Description", "chr1", 1, 1000, "+", list())
gene
}
