% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_createLncRNAGene.R
\name{createLncRNAGene}
\alias{createLncRNAGene}
\title{Create an lncRNAGene object}
\usage{
createLncRNAGene(
  ID,
  symbol,
  name,
  description,
  chr,
  start,
  end,
  strand,
  product,
  lncRNAID,
  RNASequence
)
}
\arguments{
\item{ID}{integer The gene ID (e.g., Ensembl ID or NCBI gene ID).}

\item{symbol}{character. The gene symbol.}

\item{name}{character. The gene name.}

\item{description}{character. The gene description.}

\item{chr}{character. The chromosome location.}

\item{start}{numeric. The start position of the gene.}

\item{end}{numeric. The end position of the gene.}

\item{strand}{character. The strand of the gene.}

\item{product}{list. The gene product.}

\item{lncRNAID}{character. The lncRNA ID.}

\item{RNASequence}{character. The RNA sequence.}
}
\value{
An lncRNAGene object.
}
\description{
Create an lncRNAGene object
}
\examples{
lncrna_gene <- createLncRNAGene(
   2L, "SYMBOL_LNC", "LncRNA Name",
  "LncRNA Description", "chr1", 1, 1000, "+",
  list(), "lncrna1", "RNA_SEQUENCE"
)
lncrna_gene
}
