% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_createMicroRNAGene.R
\name{createMicroRNAGene}
\alias{createMicroRNAGene}
\title{Create a microRNAGene object}
\usage{
createMicroRNAGene(
  ID,
  symbol,
  name,
  description,
  chr,
  start,
  end,
  strand,
  product,
  microRNAID,
  microRNASequence
)
}
\arguments{
\item{ID}{integer The gene ID (e.g., Ensembl ID or NCBI gene ID).}

\item{symbol}{character. The gene symbol.}

\item{name}{character. The gene name.}

\item{description}{character. The gene description.}

\item{chr}{character. The chromosome location.}

\item{start}{numeric. The start position of the gene.}

\item{end}{numeric. The end position of the gene.}

\item{strand}{character. The strand of the gene.}

\item{product}{list. The gene product.}

\item{microRNAID}{character. The microRNA ID.}

\item{microRNASequence}{character. The microRNA sequence.}
}
\value{
A microRNAGene object.
}
\description{
Create a microRNAGene object
}
\examples{
mirna_gene <- createMicroRNAGene(
  3L, "SYMBOL_MIR", "MicroRNA Name",
  "MicroRNA Description", "chr1", 1, 1000, "+",
  list(), "mirna1", "SEED_SEQ"
)
mirna_gene
}
