% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_createPiRNAGene.R
\name{createPiRNAGene}
\alias{createPiRNAGene}
\title{Create a piRNAGene object}
\usage{
createPiRNAGene(
  ID,
  symbol,
  name,
  description,
  chr,
  start,
  end,
  strand,
  product,
  piRNAID,
  piRNASequence
)
}
\arguments{
\item{ID}{integer The gene ID (e.g., Ensembl ID or NCBI gene ID).}

\item{symbol}{character. The gene symbol.}

\item{name}{character. The gene name.}

\item{description}{character. The gene description.}

\item{chr}{character. The chromosome location.}

\item{start}{numeric. The start position of the gene.}

\item{end}{numeric. The end position of the gene.}

\item{strand}{character. The strand of the gene.}

\item{product}{list. The gene product.}

\item{piRNAID}{character. The piRNA ID.}

\item{piRNASequence}{character. The piRNA sequence.}
}
\value{
A piRNAGene object.
}
\description{
Create a piRNAGene object
}
\examples{
pirna_gene <- createPiRNAGene(
  8L, "SYMBOL_PI", "piRNA Name", "piRNA Description",
  "chr1", 1, 1000, "+", list(), "pirna1", "PIRNA_SEQ"
)
pirna_gene
}
