% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_createProteinCodingGene.R
\name{createProteinCodingGene}
\alias{createProteinCodingGene}
\title{Create a ProteinCodingGene object}
\usage{
createProteinCodingGene(
  ID,
  symbol,
  name,
  description,
  chr,
  start,
  end,
  strand,
  product,
  proteinID,
  proteinSequence
)
}
\arguments{
\item{ID}{integer The gene ID (e.g., Ensembl ID or NCBI gene ID).}

\item{symbol}{character. The gene symbol.}

\item{name}{character. The gene name.}

\item{description}{character. The gene description.}

\item{chr}{character. The chromosome location.}

\item{start}{numeric. The start position of the gene.}

\item{end}{numeric. The end position of the gene.}

\item{strand}{character. The strand of the gene.}

\item{product}{list. The gene product.}

\item{proteinID}{character. The protein ID.}

\item{proteinSequence}{character. The protein sequence.}
}
\value{
A ProteinCodingGene object.
}
\description{
This function creates a new ProteinCodingGene object.
}
\examples{
protein_gene <- createProteinCodingGene(
  1L, "SYMBOL", "Gene Name", "Description",
  "chr1", 1, 1000, "+", list(), "protein1", "SEQUENCE"
)
protein_gene
}
