% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_createRRNAGene.R
\name{createRRNAGene}
\alias{createRRNAGene}
\title{Create an rRNAGene object}
\usage{
createRRNAGene(
  ID,
  symbol,
  name,
  description,
  chr,
  start,
  end,
  strand,
  product,
  rRNAID,
  rRNASequence
)
}
\arguments{
\item{ID}{integer The gene ID (e.g., Ensembl ID or NCBI gene ID).}

\item{symbol}{character. The gene symbol.}

\item{name}{character. The gene name.}

\item{description}{character. The gene description.}

\item{chr}{character. The chromosome location.}

\item{start}{numeric. The start position of the gene.}

\item{end}{numeric. The end position of the gene.}

\item{strand}{character. The strand of the gene.}

\item{product}{list. The gene product.}

\item{rRNAID}{character. The rRNA ID.}

\item{rRNASequence}{character. The rRNA sequence.}
}
\value{
An rRNAGene object.
}
\description{
Create an rRNAGene object
}
\examples{
rrna_gene <- createRRNAGene(
  7L, "SYMBOL_R", "rRNA Name", "rRNA Description",
  "chr1", 1, 1000, "+", list(), "rrna1", "RRNA_SEQ"
)
rrna_gene
}
