% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_createTRNAGene.R
\name{createTRNAGene}
\alias{createTRNAGene}
\title{Create a tRNAGene object}
\usage{
createTRNAGene(
  ID,
  symbol,
  name,
  description,
  chr,
  start,
  end,
  strand,
  product,
  tRNAID,
  tRNASequence
)
}
\arguments{
\item{ID}{integer The gene ID.}

\item{symbol}{character. The gene symbol.}

\item{name}{character. The gene name.}

\item{description}{character. The gene description.}

\item{chr}{character. The chromosome location.}

\item{start}{numeric. The start position of the gene.}

\item{end}{numeric. The end position of the gene.}

\item{strand}{character. The strand of the gene.}

\item{product}{list. The gene product.}

\item{tRNAID}{character. The tRNA ID.}

\item{tRNASequence}{character. The tRNA sequence.}
}
\value{
A tRNAGene object.
}
\description{
Create a tRNAGene object
}
\examples{
trna_gene <- createTRNAGene(
  6L, "SYMBOL_T", "tRNA Name",
  "tRNA Description", "chr1", 1, 1000, "+",
  list(), "trna1", "TRNA_SEQ"
)
trna_gene
}
