% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_Gene.R
\docType{class}
\name{Gene-class}
\alias{Gene-class}
\title{Define the Gene class
Gene class}
\description{
A class representing a generic gene.
}
\section{Slots}{

\describe{
\item{\code{ID}}{integer The gene ID (e.g., Ensembl ID or NCBI gene ID).}

\item{\code{symbol}}{character. The gene symbol.}

\item{\code{name}}{character. The gene name.}

\item{\code{description}}{character. The gene description.}

\item{\code{structure}}{GRanges. The gene structure.}

\item{\code{product}}{list. The gene product.}
}}

\examples{
gr <- GenomicRanges::GRanges(
  seqnames = "chr1",
  ranges = IRanges::IRanges(start = 1, end = 1000)
)
gene <- new("Gene",
  ID = 1L,
  symbol = "SYMBOL",
  name = "Gene Name",
  description = "Description",
  structure = gr,
  product = list()
)
}
