% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_getID.R
\name{getID}
\alias{getID}
\alias{getID,Gene-method}
\title{Get the gene ID}
\usage{
getID(object)

\S4method{getID}{Gene}(object)
}
\arguments{
\item{object}{A Gene object.}
}
\value{
The gene ID.
}
\description{
This function retrieves the ID of a gene.
}
\examples{
gene <- createGene(1L, "SYMBOL", "Gene Name", "Description", "chr1", 1, 1000, "+", list())
getID(gene)
}
