% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_getLncRNAID.R
\name{getLncRNAID}
\alias{getLncRNAID}
\alias{getLncRNAID,lncRNAGene-method}
\title{Get the lncRNA ID}
\usage{
getLncRNAID(object)

\S4method{getLncRNAID}{lncRNAGene}(object)
}
\arguments{
\item{object}{A lncRNAGene object.}
}
\value{
The lncRNA ID.
}
\description{
This function retrieves the ID of a lncRNA gene.
}
\examples{
lncrna_gene <- createLncRNAGene(
  2L, "SYMBOL_LNC", "LncRNA Name",
  "LncRNA Description", "chr1", 1, 1000, "+",
  list(), "lncrna1", "RNA_SEQUENCE"
)
getLncRNAID(lncrna_gene)
}
