% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_getMicroRNAID.R
\name{getMicroRNAID}
\alias{getMicroRNAID}
\alias{getMicroRNAID,microRNAGene-method}
\title{Get the microRNA ID}
\usage{
getMicroRNAID(object)

\S4method{getMicroRNAID}{microRNAGene}(object)
}
\arguments{
\item{object}{A microRNAGene object.}
}
\value{
The microRNA ID.
}
\description{
This function retrieves the ID of a microRNA gene.
}
\examples{
mirna_gene <- createMicroRNAGene(
  3L, "SYMBOL_MIR", "MicroRNA Name",
  "MicroRNA Description", "chr1", 1, 1000,
  "+", list(), "mirna1", "SEED_SEQ"
)
getMicroRNAID(mirna_gene)
}
