% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_getName.R
\name{getName}
\alias{getName}
\alias{getName,Gene-method}
\title{Get the gene name}
\usage{
getName(object)

\S4method{getName}{Gene}(object)
}
\arguments{
\item{object}{A Gene object.}
}
\value{
The gene name.
}
\description{
This function retrieves the name of a gene.
}
\examples{
gene <- createGene(1L, "SYMBOL", "Gene Name", "Description", "chr1", 1, 1000, "+", list())
getName(gene)
}
