% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_getPiRNASequence.R
\name{getPiRNASequence}
\alias{getPiRNASequence}
\alias{getPiRNASequence,piRNAGene-method}
\title{Get the piRNA sequence}
\usage{
getPiRNASequence(object)

\S4method{getPiRNASequence}{piRNAGene}(object)
}
\arguments{
\item{object}{A piRNAGene object.}
}
\value{
The piRNA sequence.
}
\description{
This function retrieves the sequence of a piRNA gene.
}
\examples{
pirna_gene <- createPiRNAGene(
  8L, "SYMBOL_PI", "piRNA Name",
  "piRNA Description", "chr1", 1, 1000, "+",
  list(), "pirna1", "PIRNA_SEQ"
)
getPiRNASequence(pirna_gene)
}
