% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_getProteinSequence.R
\name{getProteinSequence}
\alias{getProteinSequence}
\alias{getProteinSequence,ProteinCodingGene-method}
\title{Get the protein sequence}
\usage{
getProteinSequence(object)

\S4method{getProteinSequence}{ProteinCodingGene}(object)
}
\arguments{
\item{object}{A ProteinCodingGene object.}
}
\value{
The protein sequence.
}
\description{
This function retrieves the sequence of a protein-coding gene.
}
\examples{
protein_gene <- createProteinCodingGene(
  1L, "SYMBOL", "Gene Name",
  "Description", "chr1", 1, 1000, "+",
  list(), "protein1", "SEQUENCE"
)
getProteinSequence(protein_gene)
}
