% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_getRNASequence.R
\name{getRNASequence}
\alias{getRNASequence}
\alias{getRNASequence,lncRNAGene-method}
\title{Get the RNA sequence}
\usage{
getRNASequence(object)

\S4method{getRNASequence}{lncRNAGene}(object)
}
\arguments{
\item{object}{A lncRNAGene object.}
}
\value{
The RNA sequence.
}
\description{
This function retrieves the RNA sequence of a lncRNA gene.
}
\examples{
lncrna_gene <- createLncRNAGene(
  2L, "SYMBOL_LNC", "LncRNA Name",
  "LncRNA Description", "chr1", 1, 1000, "+",
  list(), "lncrna1", "RNA_SEQUENCE"
)
getRNASequence(lncrna_gene)
}
