% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_getRRNAID.R
\name{getRRNAID}
\alias{getRRNAID}
\alias{getRRNAID,rRNAGene-method}
\title{Get the rRNA ID}
\usage{
getRRNAID(object)

\S4method{getRRNAID}{rRNAGene}(object)
}
\arguments{
\item{object}{A rRNAGene object.}
}
\value{
The rRNA ID.
}
\description{
This function retrieves the ID of a rRNA gene.
}
\examples{
rrna_gene <- createRRNAGene(
  7L, "SYMBOL_R", "rRNA Name",
  "rRNA Description", "chr1", 1, 1000, "+", list(),
  "rrna1", "RRNA_SEQ"
)
getRRNAID(rrna_gene)
}
