% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_getSiRNAID.R
\name{getSiRNAID}
\alias{getSiRNAID}
\alias{getSiRNAID,siRNAGene-method}
\title{Get the siRNA ID}
\usage{
getSiRNAID(object)

\S4method{getSiRNAID}{siRNAGene}(object)
}
\arguments{
\item{object}{A siRNAGene object.}
}
\value{
The siRNA ID.
}
\description{
This function retrieves the ID of a siRNA gene.
}
\examples{
sirna_gene <- createSiRNAGene(
  4L, "SYMBOL_SI", "siRNA Name",
  "siRNA Description", "chr1", 1, 1000, "+",
  list(), "sirna1", "SIRNA_SEQ"
)
getSiRNAID(sirna_gene)
}
