% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_getSiRNASequence.R
\name{getSiRNASequence}
\alias{getSiRNASequence}
\alias{getSiRNASequence,siRNAGene-method}
\title{Get the siRNA sequence}
\usage{
getSiRNASequence(object)

\S4method{getSiRNASequence}{siRNAGene}(object)
}
\arguments{
\item{object}{A siRNAGene object.}
}
\value{
The siRNA sequence.
}
\description{
This function retrieves the sequence of a siRNA gene.
}
\examples{
sirna_gene <- createSiRNAGene(
  4L, "SYMBOL_SI", "siRNA Name",
  "siRNA Description", "chr1", 1, 1000, "+",
  list(), "sirna1", "SIRNA_SEQ"
)
getSiRNASequence(sirna_gene)
}
