% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_getSnoRNAID.R
\name{getSnoRNAID}
\alias{getSnoRNAID}
\alias{getSnoRNAID,snoRNAGene-method}
\title{Get the snoRNA ID}
\usage{
getSnoRNAID(object)

\S4method{getSnoRNAID}{snoRNAGene}(object)
}
\arguments{
\item{object}{A snoRNAGene object.}
}
\value{
The snoRNA ID.
}
\description{
This function retrieves the ID of a snoRNA gene.
}
\examples{
snorna_gene <- createSnoRNAGene(
  5L, "SYMBOL_SNO", "snoRNA Name",
  "snoRNA Description", "chr1", 1, 1000, "+",
  list(), "snorna1", "SNORNA_SEQ"
)
getSnoRNAID(snorna_gene)
}
