% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_getSnoRNASequence.R
\name{getSnoRNASequence}
\alias{getSnoRNASequence}
\alias{getSnoRNASequence,snoRNAGene-method}
\title{Get the snoRNA sequence}
\usage{
getSnoRNASequence(object)

\S4method{getSnoRNASequence}{snoRNAGene}(object)
}
\arguments{
\item{object}{A snoRNAGene object.}
}
\value{
The snoRNA sequence.
}
\description{
This function retrieves the sequence of a snoRNA gene.
}
\examples{
snorna_gene <- createSnoRNAGene(
  5L, "SYMBOL_SNO", "snoRNA Name",
  "snoRNA Description", "chr1", 1, 1000, "+",
  list(), "snorna1", "SNORNA_SEQ"
)
getSnoRNASequence(snorna_gene)
}
