% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_getTRNAID.R
\name{getTRNAID}
\alias{getTRNAID}
\alias{getTRNAID,tRNAGene-method}
\title{Get the tRNA ID}
\usage{
getTRNAID(object)

\S4method{getTRNAID}{tRNAGene}(object)
}
\arguments{
\item{object}{A tRNAGene object.}
}
\value{
The tRNA ID.
}
\description{
This function retrieves the ID of a tRNA gene.
}
\examples{
trna_gene <- createTRNAGene(
  6L, "SYMBOL_T", "tRNA Name",
  "tRNA Description", "chr1", 1, 1000, "+",
  list(), "trna1", "TRNA_SEQ"
)
getTRNAID(trna_gene)
}
