% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_getTRNASequence.R
\name{getTRNASequence}
\alias{getTRNASequence}
\alias{getTRNASequence,tRNAGene-method}
\title{Get the tRNA sequence}
\usage{
getTRNASequence(object)

\S4method{getTRNASequence}{tRNAGene}(object)
}
\arguments{
\item{object}{A tRNAGene object.}
}
\value{
The tRNA sequence.
}
\description{
This function retrieves the sequence of a tRNA gene.
}
\examples{
trna_gene <- createTRNAGene(
  6L, "SYMBOL_T", "tRNA Name",
  "tRNA Description", "chr1", 1, 1000, "+",
  list(), "trna1", "TRNA_SEQ"
)
getTRNASequence(trna_gene)
}
