% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_lncRNAGene.R
\docType{class}
\name{lncRNAGene-class}
\alias{lncRNAGene-class}
\title{lncRNAGene class}
\description{
A class representing a long non-coding RNA gene.
}
\section{Slots}{

\describe{
\item{\code{lncRNAID}}{character The lncRNA ID.}

\item{\code{RNASequence}}{character. The RNA sequence.}
}}

\examples{
gr <- GenomicRanges::GRanges(
  seqnames = "chr1",
  ranges = IRanges::IRanges(start = 1, end = 1000)
)
lncrna_gene <- new("lncRNAGene",
  ID = 2L, symbol = "SYMBOL_LNC",
  name = "LncRNA Name", description = "LncRNA Description",
  structure = gr, product = list(), lncRNAID = "lncrna1",
  RNASequence = "RNA_SEQUENCE"
)
}
