% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_piRNAGene.R
\docType{class}
\name{piRNAGene-class}
\alias{piRNAGene-class}
\title{piRNAGene class}
\description{
A class representing a piRNA gene.
}
\section{Slots}{

\describe{
\item{\code{piRNAID}}{character. The piRNA ID.}

\item{\code{piRNASequence}}{character. The piRNA sequence.}
}}

\examples{
gr <- GenomicRanges::GRanges(
  seqnames = "chr1",
  ranges = IRanges::IRanges(start = 1, end = 1000)
)
pirna_gene <- new("piRNAGene",
  ID = 8L, symbol = "SYMBOL_PI",
  name = "piRNA Name", description = "piRNA Description",
  structure = gr, product = list(), piRNAID = "pirna1",
  piRNASequence = "PIRNA_SEQ"
)
}
