% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_rRNAGene.R
\docType{class}
\name{rRNAGene-class}
\alias{rRNAGene-class}
\title{rRNAGene class}
\description{
A class representing a rRNA gene.
}
\section{Slots}{

\describe{
\item{\code{rRNAID}}{character. The rRNA ID.}

\item{\code{rRNASequence}}{character. The rRNA sequence.}
}}

\examples{
gr <- GenomicRanges::GRanges(
  seqnames = "chr1",
  ranges = IRanges::IRanges(start = 1, end = 1000)
)
rrna_gene <- new("rRNAGene",
  ID = 7L, symbol = "SYMBOL_R",
  name = "rRNA Name", description = "rRNA Description",
  structure = gr, product = list(), rRNAID = "rrna1",
  rRNASequence = "RRNA_SEQ"
)
}
