% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_setID.R
\name{setID}
\alias{setID}
\alias{setID<-}
\alias{setID<-,Gene-method}
\title{Set the gene ID}
\usage{
setID(object) <- value

\S4method{setID}{Gene}(object) <- value
}
\arguments{
\item{object}{A Gene object.}

\item{value}{The new gene ID.}
}
\value{
The updated Gene object.
}
\description{
Set the gene ID
}
\details{
This method sets the ID of a Gene object.
}
\examples{
gene <- createGene(1L, "SYMBOL", "Gene Name", "Description", "chr1", 1, 1000, "+", list())
setID(gene) <- 2L
getID(gene)
}
