% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_setMicroRNAID.R
\name{setMicroRNAID}
\alias{setMicroRNAID}
\alias{setMicroRNAID<-}
\alias{setMicroRNAID<-,microRNAGene-method}
\title{Set the microRNA ID}
\usage{
setMicroRNAID(object) <- value

\S4method{setMicroRNAID}{microRNAGene}(object) <- value
}
\arguments{
\item{object}{A microRNAGene object.}

\item{value}{The new microRNA ID.}
}
\value{
The updated microRNAGene object.
}
\description{
Set the microRNA ID
}
\details{
This method sets the microRNA ID of a microRNAGene object.
}
\examples{
mirna_gene <- createMicroRNAGene(
  3L, "SYMBOL_MIR", "MicroRNA Name",
  "MicroRNA Description", "chr1", 1, 1000,
  "+", list(), "mirna1", "SEED_SEQ"
)
setMicroRNAID(mirna_gene) <- "new_mirna1"
getMicroRNAID(mirna_gene)
}
