% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_setMicroRNASequence.R
\name{setMicroRNASequence}
\alias{setMicroRNASequence}
\alias{setMicroRNASequence<-}
\alias{setMicroRNASequence<-,microRNAGene-method}
\title{Set the microRNA sequence}
\usage{
setMicroRNASequence(object) <- value

\S4method{setMicroRNASequence}{microRNAGene}(object) <- value
}
\arguments{
\item{object}{A microRNAGene object.}

\item{value}{The new microRNA sequence.}
}
\value{
The updated microRNAGene object.
}
\description{
Set the microRNA sequence
}
\details{
This method sets the microRNA sequence of a microRNAGene object.
}
\examples{
mirna_gene <- createMicroRNAGene(
  3L, "SYMBOL_MIR", "MicroRNA Name",
  "MicroRNA Description", "chr1", 1, 1000, "+",
  list(), "mirna1", "SEED_SEQ"
)
setMicroRNASequence(mirna_gene) <- "NEW_SEED_SEQ"
getMicroRNASequence(mirna_gene)
}
