% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_setName.R
\name{setName}
\alias{setName}
\alias{setName<-}
\alias{setName<-,Gene-method}
\title{Set the gene name}
\usage{
setName(object) <- value

\S4method{setName}{Gene}(object) <- value
}
\arguments{
\item{object}{A Gene object.}

\item{value}{The new gene name.}
}
\value{
The updated Gene object.
}
\description{
Set the gene name
}
\details{
This method sets the name of a Gene object.
}
\examples{
gene <- createGene(
  1L, "SYMBOL", "Gene Name", "Description",
  "chr1", 1, 1000, "+", list()
)
setName(gene) <- "New Gene Name"
getName(gene)
}
