% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_setPiRNAID.R
\name{setPiRNAID}
\alias{setPiRNAID}
\alias{setPiRNAID<-}
\alias{setPiRNAID<-,piRNAGene-method}
\title{Set the piRNA ID}
\usage{
setPiRNAID(object) <- value

\S4method{setPiRNAID}{piRNAGene}(object) <- value
}
\arguments{
\item{object}{A piRNAGene object.}

\item{value}{The new piRNA ID.}
}
\value{
The updated piRNAGene object.
}
\description{
Set the piRNA ID
}
\details{
This method sets the piRNA ID of a piRNAGene object.
}
\examples{
pirna_gene <- createPiRNAGene(
  8L, "SYMBOL_PI", "piRNA Name",
  "piRNA Description", "chr1", 1, 1000,
  "+", list(), "pirna1", "PIRNA_SEQ"
)
setPiRNAID(pirna_gene) <- "new_pirna1"
getPiRNAID(pirna_gene)
}
