% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_setProduct.R
\name{setProduct}
\alias{setProduct}
\alias{setProduct<-}
\alias{setProduct<-,Gene-method}
\title{Set the gene product}
\usage{
setProduct(object) <- value

\S4method{setProduct}{Gene}(object) <- value
}
\arguments{
\item{object}{A Gene object.}

\item{value}{The new gene product.}
}
\value{
The updated Gene object.
}
\description{
Set the gene product
}
\details{
This method sets the product of a Gene object.
}
\examples{
gene <- createGene(
  1L, "SYMBOL", "Gene Name", "Description",
  "chr1", 1, 1000, "+", list()
)
setProduct(gene) <- list("New Product")
getProduct(gene)
}
