% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_setProteinID.R
\name{setProteinID}
\alias{setProteinID}
\alias{setProteinID<-}
\alias{setProteinID<-,ProteinCodingGene-method}
\title{Set the protein ID}
\usage{
setProteinID(object) <- value

\S4method{setProteinID}{ProteinCodingGene}(object) <- value
}
\arguments{
\item{object}{A ProteinCodingGene object.}

\item{value}{The new protein ID.}
}
\value{
The updated ProteinCodingGene object.
}
\description{
Set the protein ID
}
\details{
This method sets the protein ID of a ProteinCodingGene object.
}
\examples{
protein_gene <- createProteinCodingGene(
  1L, "SYMBOL", "Gene Name",
  "Description", "chr1", 1, 1000,
  "+", list(), "protein1", "SEQUENCE"
)
setProteinID(protein_gene) <- "new_protein1"
getProteinID(protein_gene)
}
