% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_setProteinSequence.R
\name{setProteinSequence}
\alias{setProteinSequence}
\alias{setProteinSequence<-}
\alias{setProteinSequence<-,ProteinCodingGene-method}
\title{Set the protein sequence}
\usage{
setProteinSequence(object) <- value

\S4method{setProteinSequence}{ProteinCodingGene}(object) <- value
}
\arguments{
\item{object}{A ProteinCodingGene object.}

\item{value}{The new protein sequence.}
}
\value{
The updated ProteinCodingGene object.
}
\description{
Set the protein sequence
}
\details{
This method sets the protein sequence of a ProteinCodingGene object.
}
\examples{
protein_gene <- createProteinCodingGene(
  1L, "SYMBOL", "Gene Name",
  "Description", "chr1", 1, 1000,
  "+", list(), "protein1", "SEQUENCE"
)
setProteinSequence(protein_gene) <- "NEW_SEQUENCE"
getProteinSequence(protein_gene)
}
