% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_setRNASequence.R
\name{setRNASequence}
\alias{setRNASequence}
\alias{setRNASequence<-}
\alias{setRNASequence<-,lncRNAGene-method}
\title{Set the RNA sequence}
\usage{
setRNASequence(object) <- value

\S4method{setRNASequence}{lncRNAGene}(object) <- value
}
\arguments{
\item{object}{A lncRNAGene object.}

\item{value}{The new RNA sequence.}
}
\value{
The updated lncRNAGene object.
}
\description{
Set the RNA sequence
}
\details{
This method sets the RNA sequence of a lncRNAGene object.
}
\examples{
lncrna_gene <- createLncRNAGene(
  2L, "SYMBOL_LNC", "LncRNA Name",
  "LncRNA Description", "chr1", 1, 1000,
  "+", list(), "lncrna1", "RNA_SEQUENCE"
)
setRNASequence(lncrna_gene) <- "NEW_RNA_SEQUENCE"
getRNASequence(lncrna_gene)
}
