% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_setRRNAID.R
\name{setRRNAID<-}
\alias{setRRNAID<-}
\alias{setRRNAID}
\alias{setRRNAID<-,rRNAGene-method}
\title{Set the rRNA ID}
\usage{
setRRNAID(object) <- value
}
\arguments{
\item{object}{A rRNAGene object.}

\item{value}{The new rRNA ID.}
}
\value{
The updated rRNAGene object.
}
\description{
Set the rRNA ID
}
\examples{
rrna_gene <- createRRNAGene(
  7L, "SYMBOL_R", "rRNA Name",
  "rRNA Description", "chr1", 1, 1000, "+",
  list(), "rrna1", "RRNA_SEQ"
)
setRRNAID(rrna_gene) <- "new_rrna1"
getRRNAID(rrna_gene)
}
