% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_setRRNASequence.R
\name{setRRNASequence<-}
\alias{setRRNASequence<-}
\alias{setRRNASequence}
\alias{setRRNASequence<-,rRNAGene-method}
\title{Set the rRNA sequence}
\usage{
setRRNASequence(object) <- value
}
\arguments{
\item{object}{A rRNAGene object.}

\item{value}{The new rRNA sequence.}
}
\value{
The updated rRNAGene object.
}
\description{
Set the rRNA sequence
}
\examples{
rrna_gene <- createRRNAGene(
  7L, "SYMBOL_R", "rRNA Name",
  "rRNA Description", "chr1", 1, 1000, "+",
  list(), "rrna1", "RRNA_SEQ"
)
setRRNASequence(rrna_gene) <- "NEW_RRNA_SEQ"
getRRNASequence(rrna_gene)
}
