% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_setSiRNAID.R
\name{setSiRNAID<-}
\alias{setSiRNAID<-}
\alias{setSiRNAID}
\alias{setSiRNAID<-,siRNAGene-method}
\title{Set the siRNA ID}
\usage{
setSiRNAID(object) <- value
}
\arguments{
\item{object}{A siRNAGene object.}

\item{value}{The new siRNA ID.}
}
\value{
The updated siRNAGene object.
}
\description{
Set the siRNA ID
}
\examples{
sirna_gene <- createSiRNAGene(
  4L, "SYMBOL_SI", "siRNA Name",
  "siRNA Description", "chr1", 1, 1000, "+",
  list(), "sirna1", "SIRNA_SEQ"
)
setSiRNAID(sirna_gene) <- "new_sirna1"
getSiRNAID(sirna_gene)
}
