% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_setSnoRNAID.R
\name{setSnoRNAID<-}
\alias{setSnoRNAID<-}
\alias{setSnoRNAID}
\alias{setSnoRNAID<-,snoRNAGene-method}
\title{Set the snoRNA ID}
\usage{
setSnoRNAID(object) <- value
}
\arguments{
\item{object}{A snoRNAGene object.}

\item{value}{The new snoRNA ID.}
}
\value{
The updated snoRNAGene object.
}
\description{
Set the snoRNA ID
}
\examples{
snorna_gene <- createSnoRNAGene(
  5L, "SYMBOL_SNO", "snoRNA Name",
  "snoRNA Description", "chr1", 1, 1000, "+",
  list(), "snorna1", "SNORNA_SEQ"
)
setSnoRNAID(snorna_gene) <- "new_snorna1"
getSnoRNAID(snorna_gene)
}
