% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_setStructure.R
\name{setStructure<-}
\alias{setStructure<-}
\alias{setStructure}
\alias{setStructure<-,Gene-method}
\title{Set the gene structure}
\usage{
setStructure(object) <- value
}
\arguments{
\item{object}{A Gene object.}

\item{value}{The new gene structure.}
}
\value{
The updated Gene object.
}
\description{
Set the gene structure
}
\examples{
gene <- createGene(
  1L, "SYMBOL", "Gene Name", "Description", "chr1",
  1, 1000, "+", list()
)
new_structure <- GenomicRanges::GRanges(
  seqnames = "chr2",
  ranges = IRanges::IRanges(
    start =
      100, end = 2000
  ), strand = "-"
)
setStructure(gene) <- new_structure
getStructure(gene)
}
