% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_setSymbol.R
\name{setSymbol}
\alias{setSymbol}
\alias{setSymbol<-}
\alias{setSymbol<-,Gene-method}
\title{Set the gene symbol}
\usage{
setSymbol(object) <- value

\S4method{setSymbol}{Gene}(object) <- value
}
\arguments{
\item{object}{A Gene object.}

\item{value}{The new gene symbol.}
}
\value{
The updated Gene object.
}
\description{
Set the gene symbol
}
\details{
This method sets the symbol of a Gene object.
}
\examples{
gene <- createGene(
  1L, "SYMBOL", "Gene Name", "Description",
  "chr1", 1, 1000, "+", list()
)
setSymbol(gene) <- "NewSymbol"
getSymbol(gene)
}
