% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_siRNAGene.R
\docType{class}
\name{siRNAGene-class}
\alias{siRNAGene-class}
\title{siRNAGene class}
\description{
A class representing a siRNA gene.
}
\section{Slots}{

\describe{
\item{\code{siRNAID}}{character. The siRNA ID.}

\item{\code{siRNASequence}}{character. The siRNA sequence.}
}}

\examples{
gr <- GenomicRanges::GRanges(
  seqnames = "chr1",
  ranges = IRanges::IRanges(start = 1, end = 1000)
)
sirna_gene <- new("siRNAGene",
  ID = 4L, symbol = "SYMBOL_SI",
  name = "siRNA Name", description = "siRNA Description",
  structure = gr, product = list(),
  siRNAID = "sirna1", siRNASequence = "SIRNA_SEQ"
)
}
