% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_snoRNAGene.R
\docType{class}
\name{snoRNAGene-class}
\alias{snoRNAGene-class}
\title{snoRNAGene class}
\description{
A class representing a snoRNA gene.
}
\section{Slots}{

\describe{
\item{\code{snoRNAID}}{character. The snoRNA ID.}

\item{\code{snoRNASequence}}{character. The snoRNA sequence.}
}}

\examples{
gr <- GenomicRanges::GRanges(
  seqnames = "chr1",
  ranges = IRanges::IRanges(start = 1, end = 1000)
)
snorna_gene <- new("snoRNAGene",
  ID = 5L, symbol = "SYMBOL_SNO",
  name = "snoRNA Name", description = "snoRNA Description",
  structure = gr, product = list(),
  snoRNAID = "snorna1", snoRNASequence = "SNO_SEQ"
)
}
