% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_summary-microRNAGene.R
\name{summary-microRNAGene}
\alias{summary-microRNAGene}
\alias{summary,microRNAGene-method}
\title{Summary Method for microRNAGene Class}
\usage{
\S4method{summary}{microRNAGene}(object)
}
\arguments{
\item{object}{A microRNAGene object.}
}
\value{
A summary of the microRNA gene information.
}
\description{
Summary method for microRNAGene objects
}
\details{
Provides a summary of the microRNA gene information.
}
\examples{
mirna_gene <- createMicroRNAGene(
  3L, "SYMBOL_MIR", "MicroRNA Name",
  "MicroRNA Description", "chr1", 1, 1000,
  "+", list(), "mirna1", "SEED_SEQ"
)
summary(mirna_gene)
}
