% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_summary-piRNAGene.R
\name{summary-piRNAGene}
\alias{summary-piRNAGene}
\alias{summary,piRNAGene-method}
\title{Summary Method for piRNAGene Class}
\usage{
\S4method{summary}{piRNAGene}(object)
}
\arguments{
\item{object}{A piRNAGene object.}
}
\value{
A summary of the piRNA gene information.
}
\description{
Summary method for piRNAGene objects
}
\details{
Provides a summary of the piRNA gene information.
}
\examples{
pirna_gene <- createPiRNAGene(
  8L, "SYMBOL_PI", "piRNA Name",
  "piRNA Description", "chr1", 1, 1000,
  "+", list(), "pirna1", "PIRNA_SEQ"
)
summary(pirna_gene)
}
