% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_summary-rRNAGene.R
\name{summary-rRNAGene}
\alias{summary-rRNAGene}
\alias{summary,rRNAGene-method}
\title{Summary Method for rRNAGene Class}
\usage{
\S4method{summary}{rRNAGene}(object)
}
\arguments{
\item{object}{A rRNAGene object.}
}
\value{
A summary of the rRNA gene information.
}
\description{
Summary method for rRNAGene objects
}
\details{
Provides a summary of the rRNA gene information.
}
\examples{
rrna_gene <- createRRNAGene(
  7L, "SYMBOL_R", "rRNA Name",
  "rRNA Description", "chr1", 1, 1000,
  "+", list(), "rrna1", "RRNA_SEQ"
)
summary(rrna_gene)
}
