% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_summary-siRNAGene.R
\name{summary-siRNAGene}
\alias{summary-siRNAGene}
\alias{summary,siRNAGene-method}
\title{Summary Method for siRNAGene Class}
\usage{
\S4method{summary}{siRNAGene}(object)
}
\arguments{
\item{object}{A siRNAGene object.}
}
\value{
A summary of the siRNA gene information.
}
\description{
Summary method for siRNAGene objects
}
\details{
Provides a summary of the siRNA gene information.
}
\examples{
sirna_gene <- createSiRNAGene(
  4L, "SYMBOL_SI", "siRNA Name",
  "siRNA Description", "chr1", 1, 1000,
  "+", list(), "sirna1", "SIRNA_SEQ"
)
summary(sirna_gene)
}
