% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_summary-tRNAGene.R
\name{summary-tRNAGene}
\alias{summary-tRNAGene}
\alias{summary,tRNAGene-method}
\title{Summary Method for tRNAGene Class}
\usage{
\S4method{summary}{tRNAGene}(object)
}
\arguments{
\item{object}{A tRNAGene object.}
}
\value{
A summary of the tRNA gene information.
}
\description{
Summary method for tRNAGene objects
}
\details{
Provides a summary of the tRNA gene information.
}
\examples{
trna_gene <- createTRNAGene(
  6L, "SYMBOL_T", "tRNA Name",
  "tRNA Description", "chr1", 1, 1000,
  "+", list(), "trna1", "TRNA_SEQ"
)
summary(trna_gene)
}
